Huge bunch of changes.
This commit contains a bit of UI toying, but more importantly, it also contains
a first implementation of the caching model. LocalFolderPhotoSource and
LocalFilePhoto are sources describing a very dumb set of local files. These are
cached in the MainCachePhotoSource, which uses CachePhotoSource and CachePhoto
objects to store this cache info. Still misses a ton, but storing it to the
database works.

Also added TripodSqliteConnection, a wrapper around HyenaSqliteConnection,
which toggles some of the flags to avoid sqlite slowness (takes the insert
of 260 rows from 27 seconds to .7 seconds on my slow netbook harddisk).