First stab at awesome mipmap caching.
One of the ideas I'd like to try out for making image display lighting-fast is
by generating a mipmap [1] cache of all images. Thumbnails and smaller versions
of the photo can then be pulled from this cache, which contains only stripped
down JPEGs. The full file can be loaded in the background, but it is really
only needed when zooming in: the largest cached copy should be able to fill the
screen.

This is especially important for RAW files or files with processing actions,
they take a long time to load. The cache hides this.

Currently, with a maximum tile size of 1600x1200, this amounts to around 330 kb
per photo. For my 20000 photo library this would mean (pessimistic estimate) a
cache of around eight gigabyte. I gladly accept that for the speed we gain.

Seems to work (fast!) but still needs unit tests.

[1] http://en.wikipedia.org/wiki/Mipmap