First implementation of a PhotoViewWidget and a ImageCache
The PhotoGridView makes use of Hyeana.Data.Gui to display a model of IPhoto.
In addition ImageSurfaceCache provides access to the cached image data of
MipMaps. Therfore, PhotoViewThumbnailCache holds a number of images in memory
and MipMaps are used as persitent cache of different image sizes on
the filesystem. A rudimentary implementation adds all together to the
MainWindow.