Add shamrock macro's for finding mono-2.pc.
The ABI changed in mono 2.7 and as such, the pkg-config file was renamed from
mono.pc to mono-2.pc. This patch adds shamrock macro's for finding the desired
mono module. As well as macro's for checking for assemblies in the specified
GAC. No 1.0 profile assembly check macro has been added for mono-2, as I think
this has been deprecated.

https://bugzilla.gnome.org/show_bug.cgi?id=614980