Automatically load smaller/larger sizes when needed.
When the scaling factor goes above or below a certain level, the grid will
automatically switch to a different mipmap size. This means that when you zoom
out a lot, only a large pile of very small mipmap pieces will be loaded, yet
when you zoom in, a high quality piece is used.