Enable selective upscaling.
When zooming in, thumbnails are upscaled if we know that the image is
larger than the thumbnail we currently have. As long as the icon size is
smaller than the largest mipmap, this means that we'll always get a
better size soon.

To do this, the size of the photo needs to be exposed through IPhoto.